+++ shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/thirdParty/at/asitplus/wallet/lib/data/CredentialScheme.kt
@@ -14,11 +14,13 @@ import at.asitplus.valera.resources.credential_scheme_icon_label_tax_id
 import at.asitplus.valera.resources.credential_scheme_label_certificate_of_residence
 import at.asitplus.valera.resources.credential_scheme_label_company_registration
 import at.asitplus.valera.resources.credential_scheme_label_eu_pid
+import at.asitplus.valera.resources.credential_scheme_label_eu_pid_sdjwt
 import at.asitplus.valera.resources.credential_scheme_label_healthid
 import at.asitplus.valera.resources.credential_scheme_label_id_austria
 import at.asitplus.valera.resources.credential_scheme_label_mdl
 import at.asitplus.valera.resources.credential_scheme_label_power_of_representation
 import at.asitplus.valera.resources.credential_scheme_label_tax_id
+import at.asitplus.valera.resources.credential_scheme_label_tax_id_2025
 import at.asitplus.wallet.companyregistration.CompanyRegistrationScheme
 import at.asitplus.wallet.cor.CertificateOfResidenceScheme
 import at.asitplus.wallet.eupid.EuPidScheme
@@ -28,6 +30,7 @@ import at.asitplus.wallet.idaustria.IdAustriaScheme
 import at.asitplus.wallet.lib.data.ConstantIndex
 import at.asitplus.wallet.mdl.MobileDrivingLicenceScheme
 import at.asitplus.wallet.por.PowerOfRepresentationScheme
+import at.asitplus.wallet.taxid.TaxId2025Scheme
 import at.asitplus.wallet.taxid.TaxIdScheme
 import data.credentials.CertificateOfResidenceCredentialAttributeTranslator
 import data.credentials.CompanyRegistrationCredentialAttributeTranslator
@@ -45,26 +48,28 @@ import org.jetbrains.compose.resources.stringResource
 fun ConstantIndex.CredentialScheme?.uiLabel(): String = when (this) {
     is IdAustriaScheme -> stringResource(Res.string.credential_scheme_label_id_austria)
     is EuPidScheme -> stringResource(Res.string.credential_scheme_label_eu_pid)
-    is EuPidSdJwtScheme -> stringResource(Res.string.credential_scheme_label_eu_pid)
+    is EuPidSdJwtScheme -> stringResource(Res.string.credential_scheme_label_eu_pid_sdjwt)
     is MobileDrivingLicenceScheme -> stringResource(Res.string.credential_scheme_label_mdl)
     is PowerOfRepresentationScheme -> stringResource(Res.string.credential_scheme_label_power_of_representation)
     is CertificateOfResidenceScheme -> stringResource(Res.string.credential_scheme_label_certificate_of_residence)
     is CompanyRegistrationScheme -> stringResource(Res.string.credential_scheme_label_company_registration)
     is HealthIdScheme -> stringResource(Res.string.credential_scheme_label_healthid)
     is TaxIdScheme -> stringResource(Res.string.credential_scheme_label_tax_id)
+    is TaxId2025Scheme -> stringResource(Res.string.credential_scheme_label_tax_id_2025)
     else -> this?.identifier ?: "unknown"
 }
 
 suspend fun ConstantIndex.CredentialScheme?.uiLabelNonCompose(): String = when (this) {
     is IdAustriaScheme -> getString(Res.string.credential_scheme_label_id_austria)
     is EuPidScheme -> getString(Res.string.credential_scheme_label_eu_pid)
-    is EuPidSdJwtScheme -> getString(Res.string.credential_scheme_label_eu_pid)
+    is EuPidSdJwtScheme -> getString(Res.string.credential_scheme_label_eu_pid_sdjwt)
     is MobileDrivingLicenceScheme -> getString(Res.string.credential_scheme_label_mdl)
     is PowerOfRepresentationScheme -> getString(Res.string.credential_scheme_label_power_of_representation)
     is CertificateOfResidenceScheme -> getString(Res.string.credential_scheme_label_certificate_of_residence)
     is CompanyRegistrationScheme -> getString(Res.string.credential_scheme_label_company_registration)
     is HealthIdScheme -> getString(Res.string.credential_scheme_label_healthid)
     is TaxIdScheme -> getString(Res.string.credential_scheme_label_tax_id)
+    is TaxId2025Scheme -> getString(Res.string.credential_scheme_label_tax_id_2025)
     else -> this?.identifier ?: "unknown"
 }
 
@@ -79,6 +84,7 @@ fun ConstantIndex.CredentialScheme?.iconLabel(): String = when (this) {
     is CompanyRegistrationScheme -> stringResource(Res.string.credential_scheme_icon_label_company_registration)
     is HealthIdScheme -> stringResource(Res.string.credential_scheme_icon_label_healthid)
     is TaxIdScheme -> stringResource(Res.string.credential_scheme_icon_label_tax_id)
+    is TaxId2025Scheme -> stringResource(Res.string.credential_scheme_icon_label_tax_id)
     else -> this?.identifier ?: "unknown"
 }
 
@@ -92,5 +98,6 @@ fun ConstantIndex.CredentialScheme.getLocalization(path: NormalizedJsonPath): St
     is EuPidSdJwtScheme -> { EuPidCredentialAttributeTranslator.translate(path) }
     is IdAustriaScheme -> { IdAustriaCredentialAttributeTranslator.translate(path) }
     is TaxIdScheme -> { TaxIdCredentialAttributeTranslator.translate(path) }
+    is TaxId2025Scheme -> { TaxIdCredentialAttributeTranslator.translate(path) }
     else -> { IdAustriaCredentialAttributeTranslator.translate(path) }
 }