+++ CHANGELOG.md
@@ -8,6 +8,8 @@
  * Add option to open URL from verifier after successful authentication
  * Add support for EU PID with SD-JWT names from ARF 1.8.0 onwards
  * Gracefully handle the case where the token status cannot be evaluated
+ * Add support for more age over claims in EU PID and mDL
+ * Update Tax ID credential, to fix `sdJwtType`
 
 # Release 5.5.2
  * Fix presentation of ISO credentials